View DataSourceView content
